added db functionality in student endpoint

+++ internal/types/types.go
@@ -1,8 +1,8 @@
 package types
 
 type Student struct {
-	Id    int
-	Name  string `validate:"required"`
+	Id    int64
+	Name  string `validate:"required"` // term: feild validation
 	Email string `validate:"required"`
 	Age   int    `validate:"required"`
 }
